requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
cloudinary==1.32.0
Django==4.1.7
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
